# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Entertainment App",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/bookmark/": {
      "get": {
        "summary": "Get User bookmarks data",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "firebase tokenid",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/bookmark/add": {
      "post": {
        "summary": "Add bookmark",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "firebase token id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "movieId": {
                    "type": "integer"
                  },
                  "mediaType": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "movieId": 508442,
                "mediaType": "movies"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/bookmark/remove": {
      "post": {
        "summary": "Remove added bookmark",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "firebase token ",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              },
              "example": "{\"movieId\": 508442, \"mediaType\": \"movies\"}"
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
...
    "securitySchemes": {}
  },
  "servers": [
    {
      "url": "https://entertainment-backend.vercel.app"
    }
  ]
}